CPLD logic, fully functional, as of now
(fix concatenation mistake)

Im not 100% convinced that the interrupt controller logic is 100% proper. Im observing some odd looking behaviour on the scope where it looks like the IPL signals do not latch all at the same time. While Im not able to produce any incorrect behaviour (as I was able to in an earlier iteration), this might still act up at some point in the future...