Assert Supervisor Data function code when ethernet is bus master
Although fixed in the CPLD for on-board memory accesses, for the sake of external memory boards that may require a function code that isnt all 1's, FC1 is pulled low to indicate a supervisor function code.

With the 3 FCx signals pulled high by default, pulling FC1 low generates the Supervisor Data function code.