Replace individual resistors for ethernet LEDs with resistor network
It wasnt clear from the LXT901 datasheet whether it would produce a delay to blink the TXD/RXD LEDs, so individual resistor footprints were provided in case some lower value resistors were needed for these LEDs. But it was found through rev 1 that the LXT901 does implement its own delay, so the resistors for all of these LEDs can be the same afterall.