Some change, not 100% sure what it was
Might not be anything in particular, just one silkscreen change that I can see in a diff.

Mothballing this design for now because the read/write logic needs a redesign to support slower speed cards since IORDY does not function the way I expected it to - it is not used per bus cycle, and may rarely if ever be used, so it cant be used to stretch 68k bus cycles to suit the card.