Changing buffer direction/enable logic
Rather than having the data bus buffers constantly enabled and only changing direction based on a read, the buffers between the expansion bus and the board are now disabled, and are enabled when the card is selected. This should keep the board "quiet" when it is not actively addressed.

GAL logic will need to be updated to reflect some new/changed pins.