Add new bootloader
This is a re-write of the original bootloader and companion script

* Interface is more consistent with other scripts I have written in recent times
* Can now identify and write to flash chips in-system

Importantly, it seems to start and run fine the first time every time. The previous bootloader had an issue where after power on the board would need to be reset before the bootloader could be interacted with, but that issue now seems to be gone... yay! Not sure exactly why though...